Use DDR0 to load TF-A and U-Boot
As DDR1 is not routed on some boards, for example ECAM03DM. RISC0 MMU is
configured to provide access to DDR0 and load TF-A and U-Boot there.

When RISC0 comes out of reset it works in error level (ERL=1) with
Fixed Mapping MMU where DDR0 (0x80000000) is not accessible. TLB-based
MMU is activated with single entry to provide access to DDR0. Final
address translation is as follows:

  kuseg: 0x00000000-0x7FFFFFFF -> 0x00000000-0x7FFFFFFF
  kseg0: 0x80000000-0x9FFFFFFF -> 0x00000000-0x1FFFFFFF
  kseg1: 0xA0000000-0xBFFFFFFF -> 0x00000000-0x1FFFFFFF
  kseg2: 0xC0000000-0xC0FFFFFF -> 0x80000000-0x80FFFFFF (DDR0, TLB)
         0xC1000000-0xDFFFFFFF invalid
  kseg3: 0xE0000000-0xFFFFFFFF invalid

Note that DDR1 is not accessible anymore.

Issue: #MCOM03SW-778
Change-Id: I82c4b26db65a61b8deba98ddfbbaa8dea9e35d49