scripts: Protect TF-A/TEE memory range
Memory range used by TF-A and TEE must not be available in U-Boot and
Linux. The same changes in the memory map have already been added to
sbl-tl package earlier.

The patch moves code, which is responsible for setting secure regions,
to the separate static library "libcommon.a". This library provides the
function "set_secure_region" which is responsible for setup of TF-A and
TEE regions in DDR memory map:

* the range 0x880000000 - 0x88FFFFFFF is removed from U-Boot/Linux
memory map;
* the range 0x880000000 - 0x88FFFFFFF is marked as secure for test
purpose in case of boot in VS_EN = 0 mode.

The library is linking to "sbl-xip-*.elf" and "set-secure-regions.elf".
The "set-secure-regions.elf" is built as part of the "scripts" project
and is used in *.mdb scripts to setup secure regions and to take them
into account in the DDR memory map. To do that the ELF file is loaded
with 'loadelf' MDB function and executed. We define the label 'jtag'
as break point in 'startup.S' to return back to MDB script and continue.

Issue: #MCOM03SW-2154
Issue: #MCOM03SW-1807
Issue: #MCOM03SW-2166
Change-Id: I06d890e1850a3c41cc3b74a52949b8d2d55d8e4d