Switch to use #pragma once
This patch refactors all header files in the project to use #pragma
once directive instead of header guards. Since macros guards must be
unique, the programmer is required to verify that every header file
fulfills this condition. As the number of header files grows, it
becomes difficult to monitor compliance with this requirement.
Using #pragma once resolves this issue. Also it reduces number of lines
in code and excludes the need to create macros by yourself according
to the coding style that can be changed.

Change-Id: Ibe6175a118fb4587fc5437647392bd6786b5e9b3