mips: Enable external interrupts in sbl-s3
The interrupt manager functionality to enable and disable external MIPS
interrupts was lost during the SBL-S3 integration into the SBL project.
As a result, TF-A does not receive any response when it sends IPC
messages to SBL. However, this disadvantage is leveled by the fact that
TF-A has its own mechanism to work with WDT. This patch adds functions
that enable and disable MIPS external interrupts by target number
during the initialization and deinitialization of the interrupt
manager, respectively. Also, all names of functions, macros, and
definitions that refer to the CP0 register are changed in compliance
with the documentation.

Change-Id: I1ad187fc3f19c663ed516e32537b9cf7374b2ff0