Fix conditional brackets
For a one-line statement in conditional blocks the brackets are not
required, but it becomes hard to track the program behavior in
situations when one part of the block consists of multiple statements,
and the other doesn't. This patch adds brackets for one-line statements
in conditional blocks in case other parts of the blocks have multiple
statements.

Change-Id: I1d48d59f713f2d545e1166e81089e46e848ad47c