Remove utils-def.h from assembler files
There is an issue in the codebase that utils-def.h file is indirectly
included into assembler files. As a result, macros and definitions in
utils-def.h file must fulfill both C and assembler coding rules and
those that do not belong to assembler must be excluded by guard
wrapping. This patch restructures files so that utils-def.h file is no
longer indirectly included into assembler files.

Change-Id: I6924e13a392fa865b2f5ba4817fdcc98cc4087c5