Change position values from masks to numbers
There is an issue in the codebase that some positional definitions have
different representations across the project. Some definitions are
declared as numbers while the others declared as masks, which could
lead to coding errors and make it impossible to access some positional
numbers directly. This commit fixes the problem by changing all
position definitions from masks to numbers.

Change-Id: I6d1154ab8193fea87e3e6c5d73b0f81449498782