crypto: Save certificate SHA checksum
The project is designed in such a way that the checksum calculation
process is performed not on the entire certificate but only on its
part, namely on tbsCertificate. As a result, if it is required to
obtain the SHA, it has to be recalculated, performing additional
procedures for skipping a part of the certificate. This patch stores
the calculated SHA in a separate field of the structure so that it may
be compared with OTP without having to do the calculations again.

Issue: #MCOM03SW-2603
Change-Id: I5a99bc3b511e8faedfa62aa01316d8c85acebfa5