env: Fix undefined behavior
According to Programming languages C, ISO/IEC 9899:1999 §7.1.4
(Use of library functions) if an argument to a function has an invalid
value (such as a value outside the domain of the function, or a pointer
outside the address space of the program, or a null pointer, or a
pointer to non-modifiable storage when the corresponding parameter
is not const-qualified), the behavior is undefined.

So let's fix passing null pointer as argument 2 to memcpy, which is
declared to never be null.

Issue: #MCOM03SW-2603
Change-Id: I9e4282f0d4b47532cfe665f7cdd58aebaf8a5939