Set LSP1 PLL and UCG default frequency
We are going to use LSP1 TIMERS0 as system timer. The LSP1 TIMERS0
frequency has to be fixed across boot stages. The PLL coefficients
and UCG dividers are chosen to the ones set in our Linux Kernel.
As PLL has to be tuned, the related clocks of peripherals used in
firmware are tuned too. The resulting frequencies are the following:

      lsperiph1_pll     614250000
        lsp1_sys_clk    87750000
        timers0_clk     87750000
        uart0_clk       26706522
        gpio1_dbclk     1050000

Issue: #MCOM03SW-3347
Change-Id: I87dfe47ec20006371bf8a47ee2bd30dc567a8e62