sbl-s2: Provide version via SBL environment
The new read-only (attribute 'so') environment variable 'sbl-s2-ver' is
introduced. The variable contains version string of SBL-S2 in the
following format:

  (<TIME>, <DATE>)-g<COMMIT SHA[-dirty]|unknown>[-<BUILD_ID>]

If it isn't possible to get commit SHA for any reason, the suffix
"-gunknown" is used.

If <BUILD_ID> isn't provided, the suffix "-<BUILD_ID>" is absented.

Also the definition of the version string is moved to the
'platform-def-common.h' to reuse it in the welcome message and
to set the environment variable 'sbl-s2-ver'.

Issue: #MCOM03SW-3765
Change-Id: I7cae9f293a5d64a31485823953b2add23be464c1