Rework WDT initialization
The patch allows WDT to be enabled using the RISC0 IPC service if
it was not started in the early boot process.

Also the common definitions to convert time units are moved from
drivers/timer/timer.h to libs/utils-def.h to reuse it in the WDT driver.

Issue: #MCOM03SW-3841
Change-Id: I0ffa0fd17d84393134b202bf1cff7075658bb769