Remove soc_debug_if_disable
RISC0 debugging with JTAG operates via risc0 oncd but the provided
soc_debug_if_disable function disables the BPAM clock, which is a part
of the UltraSoC subsystem. That is the reason why this patch deletes
the disabling SoC debug interface functionality since it is not tested
and it is impossible to disable RISC0 debugging programmatically with
it. As an alternative, OTP::fuse0.ust_debug_disable (disable debug on
UltraSoC) and OTP::fuse0.bringupdbg_disable (disable debug on JTAG)
could be burned on MCom-03R end devices in order to satisfy the
security requirements.

Change-Id: I894b033c48df497f6430bef0f4ee2803371cb6a3